feat(sendCommand.go): add support for sending commands to specific devices based on device ID
The `Command` struct has been updated to include a `DeviceID` field. The `CommandRequest` struct has also been updated to include a `DeviceID` field. This allows for specifying the device ID when sending a command.

In the `SendCommand` function, the `command` variable is now initialized with the `DeviceID` field from the `command_request`. The `topic` variable is constructed using the `DeviceID` to specify the destination of the command. The command is then published to the specified topic using MQTT.